+++ com.unity.render-pipelines.high-definition/Editor/Material/ShaderGraph/ShaderGraphVersion.cs
@@ -10,5 +10,6 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
     {
         Initial = 0,
         FirstTimeMigration = Initial,
+        ExposableProperties,
     }
 }