+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Global settings check in Wizard
 - Added localization on Wizard window
 - Added an info box for micro shadow editor (case 1322830).
+- Added expose checkbox for shadergraph sub target properties.
 
 ### Fixed
 - Fixed Intensity Multiplier not affecting realtime global illumination.